this might not work
if it doesn't, I'm just going to revert

+++ tests/test_cpp.py
@@ -88,7 +88,7 @@ class Test(unittest.TestCase):
         # as well as str, repr, and getitem methods
         a = numpy.array([1, 4, 1, 0])
         b = numpy.array([2, 9, 5, 10])
-        c = numpy.arange(10, dtype=numpy.int16)
+        c = numpy.arange(10)
         test = awkward.cpp.JaggedArray(a, b, c)
         assert repr(test) == "<JaggedArray [[1] [4 5 6 7 8] [1 2 3 4] [0 1 2 3 4 5 6 7 8 9]]>"
